requirements.txt
pip==21.2.2
colorlog==6.6.0
toml==0.10.2
zlib==1.2.11
wheel==0.37.1
brotli==1.0.9
openssl==1.1.1n
cryptography==3.4.8
tornado==6.1
typing_extensions==4.1.1
keyring==23.4.0
setuptools==58.0.4
docutils==0.18.1
tensorflow==2.4.1
numpy==1.21.2
pandas==1.3.5
scikit-learn==1.0.2
keras==2.4.3
matplotlib==3.5.1
gputil==1.4.0
psutil==5.8.0
tqdm==4.62.3
torchbnn==1.2
loguru==0.6.0
plotly==5.6.0
scipy==1.7.3
bokeh==2.4.2
optuna==2.10.0
torchvision==0.10.0
seaborn==0.11.2
tslearn==0.5.2